refactor(QChip): replace legacy QChip with the new one
