--- src/lib/components/chip/index.scss
@@ -1,105 +1,0 @@
-@use "$css/mixins";
-
-// QChip
-.q-chip {
-  box-sizing: content-box;
-  display: inline-flex;
-  align-items: center;
-  justify-content: center;
-  height: 2.5rem;
-  min-width: 2.5rem;
-  font-size: 0.875rem;
-  font-weight: 500;
-  background-color: var(--secondary);
-  color: var(--on-secondary);
-  text-transform: none;
-  border-radius: 0.5em;
-  transition:
-    var(--speed3) transform,
-    var(--speed3) border-radius,
-    var(--speed3) padding;
-  line-height: normal;
-  text-decoration: none;
-  cursor: pointer;
-  border: none;
-
-  @extend .no-select;
-  @include mixins.padding("x", "md");
-  @include mixins.margin("x");
-  @include mixins.gap("md");
-  @include mixins.btn-image("chip");
-
-  &.q-chip--sm {
-    height: 2rem;
-    min-width: 2rem;
-  }
-
-  &.q-chip--lg {
-    height: 3rem;
-    min-width: 3rem;
-  }
-
-  &.q-chip--bordered {
-    @include mixins.transparent($ripple: true);
-    @include mixins.border(var(--secondary));
-    color: var(--secondary);
-  }
-
-  &.q-chip--rounded {
-    border-radius: 1.25rem;
-
-    &.q-chip--sm {
-      border-radius: 1rem;
-    }
-
-    &.q-chip--lg {
-      border-radius: 1.5rem;
-    }
-  }
-
-  &.q-chip--filled {
-    background-color: var(--secondary-container);
-    color: var(--on-secondary-container);
-    border: none;
-  }
-
-  &.q-chip--vertical {
-    flex-direction: column;
-    height: auto;
-    max-height: none;
-
-    @include mixins.padding("y");
-    @include mixins.gap("xs");
-  }
-
-  &.q-chip[aria-disabled="true"] {
-    @extend .disabled;
-    &.q-chip--bordered {
-      border-color: var(--on-surface);
-    }
-  }
-
-  &:is(:hover, :focus):not([aria-disabled])::after {
-    content: "";
-    position: absolute;
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 100%;
-    background-color: var(--on-primary);
-  }
-
-  &:hover:not([aria-disabled])::after {
-    opacity: 0.08;
-  }
-
-  &:focus:not([aria-disabled])::after {
-    opacity: 0.16;
-  }
-}
-
-// QChip icons
-.q-chip > .q-icon {
-  --size: 1.25rem;
-  margin-inline: -0.5rem;
-}
